More fixes for https://github.com/batchatco/go-native-netcdf/issues/4
This handles odd-sized layouts and comes with a unit test.

The handling of var slices with chunked layouts is not 100% correct still needs work.
It works most of the time though.

+++ netcdf/hdf5/compound.go
@@ -140,7 +140,7 @@ func (compoundManagerType) parse(hr heapReader, c caster, attr *attribute, bitFi
 		logger.Info("compound alloced", df.Count(), df.Rem()+df.Count(),
 			"attrSize=", attrSize)
 		bff := df
-		if attrSize > df.Rem() {
+		if int64(attrSize) > df.Rem() {
 			logger.Info("Adding fill value reader")
 			// bff = makeFillValueReader(obj, df, attrSize)
 		}
